♻️ refactor: apis, queries 파일 리팩토링
